+++ app/[domain]/(default)/page.tsx
@@ -3,7 +3,7 @@ import { notFound } from 'next/navigation';
 import { getSiteData } from '$utils/fetchers';
 
 import { HomePageContent } from './content';
-import { ClaimLemonHeadCard } from '$lib/components/features/lemonheads/ClaimLemonHeadCard';
+// import { ClaimLemonHeadCard } from '$lib/components/features/lemonheads/ClaimLemonHeadCard';
 import { WhoToFollow } from '$lib/components/features/lens-account/WhoToFollow';
 import { UpcommingEventsCard } from '$lib/components/features/event/UpcommingEventsCard';
 import { LemonadeStandCard } from '$lib/components/features/LemonadeStandCard';
@@ -21,7 +21,7 @@ export default async function SiteHomePage({ params }: { params: Promise<{ domai
     <div className="flex flex-col-reverse md:grid md:grid-cols-[1fr_336px] gap-5 md:gap-[72px] items-start pb-10 mt-0 md:mt-11">
       <HomePageContent />
       <div className="flex flex-col gap-4">
-        <ClaimLemonHeadCard />
+        {/* <ClaimLemonHeadCard /> */}
         <LemonadeStandCard />
         <UpcommingEventsCard />
         <WhoToFollow />